edycja notification na sweetAlert

+++ frontend/TattooMe/src/app/component/favorite/favorite.component.ts
@@ -3,6 +3,7 @@ import { FavoriteService } from '../../service/favorite.service';
 import { Router } from '@angular/router';
 import { FavoriteArtist } from '../../model/favorite-artist';
 import { NotificationService } from '../../service/notification.service';
+import Swal from 'sweetalert2';
 
 
 @Component({
@@ -30,20 +31,35 @@ export class FavoriteComponent implements OnInit {
         this.favorites = data;
       },
       error: (err) => {
-        this.notification.showError("Nie udało się załadować ulubionych artystów", err);
+        console.log(err)
       },
     });
   }
 
   removeFavorite(artistId: string): void {
-    this.favoriteService.removeFavorite(artistId).subscribe({
-      next: () => {
-        this.loadFavorites();
-        this.notification.showSuccess("Artysta został usunięty z ulubionych");
-      },
-      error: (err) => {
-        this.notification.showError("Nie udało się usunąć artysty z ulubionych", err);
-      },
+    Swal.fire({
+      title: 'Usunąć artystę z ulubionych?',
+      text: 'Tej akcji nie można cofnąć.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      background: '#1e1e1e',
+      color: '#ffffff',
+      confirmButtonText: 'Tak, usuń',
+      cancelButtonText: 'Anuluj'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.favoriteService.removeFavorite(artistId).subscribe({
+          next: () => {
+            this.loadFavorites();
+            this.notification.showSuccess("Artysta został usunięty z ulubionych");
+          },
+          error: (err) => {
+            this.notification.showError("Nie udało się usunąć artysty z ulubionych");
+          }
+        });
+      }
     });
   }
