+++ frontend/TattooMe/src/app/component/profile/portfolio/portfolio.component.ts
@@ -2,6 +2,7 @@ import { Component, Input, OnInit } from '@angular/core';
 import { Portfolio } from '../../../model/portfolio';
 import { PortfolioService } from '../../../service/portfolio.service';
 import { NotificationService } from '../../../service/notification.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-portfolio',
@@ -17,7 +18,7 @@ export class PortfolioComponent implements OnInit {
   selectedFile: File | null = null;
   showAllPortfolio = false;
 
-  constructor(private portfolioService: PortfolioService, private notification: NotificationService) {}
+  constructor(private portfolioService: PortfolioService, private notification: NotificationService) { }
 
   ngOnInit(): void {
     this.loadPortfolio();
@@ -31,12 +32,29 @@ export class PortfolioComponent implements OnInit {
   }
 
   deleteImagePortfolio(id: string): void {
-    this.portfolioService.delete(id).subscribe({
-      next: () => {
-        this.portfolioItems = this.portfolioItems.filter(p => p.id !== id);
-        this.notification.showSuccess("Zdjęcie zostało usunięte");
-      },
-      error: err => this.notification.showError("Nie udalo sie usunac zdjecia", err)
+    Swal.fire({
+      title: 'Usunąć to zdjęcie z portfolio?',
+      text: 'Tej akcji nie można cofnąć.',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#d33',
+      cancelButtonColor: '#3085d6',
+      background: '#1e1e1e',
+      color: '#ffffff',
+      confirmButtonText: 'Tak, usuń',
+      cancelButtonText: 'Anuluj'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.portfolioService.delete(id).subscribe({
+          next: () => {
+            this.portfolioItems = this.portfolioItems.filter(p => p.id !== id);
+            this.notification.showSuccess("Zdjęcie zostało usunięte");
+          },
+          error: (err) => {
+            this.notification.showError("Nie udało się usunąć zdjęcia");
+          }
+        });
+      }
     });
   }
 
@@ -67,18 +85,18 @@ export class PortfolioComponent implements OnInit {
   }
 
   toggleFeatured(item: Portfolio): void {
-  if (!item) return;
+    if (!item) return;
 
-  const newState = !item.featured;
+    const newState = !item.featured;
 
-  this.portfolioService.updateFeatured(this.userId, item.id, newState)
-    .subscribe({
-      next: () => {
-        item.featured = newState;
-      },
-      error: (err) => {
-        console.error('Nie udało się zmienić stanu featured:', err);
-      }
-    });
-}
+    this.portfolioService.updateFeatured(this.userId, item.id, newState)
+      .subscribe({
+        next: () => {
+          item.featured = newState;
+        },
+        error: (err) => {
+          console.error('Nie udało się zmienić stanu featured:', err);
+        }
+      });
+  }
 }
